Move BillingProjectAudit view to new app
